+++ video_xblock/tests/acceptance/__init__.py
@@ -1,0 +1,3 @@
+"""
+Acceptance test for VideoXBlock.
+"""